refactor: say page strutrue
