+++ src/app/(app)/(note-topic)/notes/(topic-detail)/topics/[slug]/page.tsx
@@ -7,8 +7,10 @@ import { useParams } from 'next/navigation'
 import { LoadMoreIndicator } from '~/components/modules/shared/LoadMoreIndicator'
 import { TimelineList } from '~/components/ui/list/TimelineList'
 import { Loading } from '~/components/ui/loading'
-import { BottomToUpSoftScaleTransitionView } from '~/components/ui/transition/BottomToUpSoftScaleTransitionView'
-import { BottomToUpTransitionView } from '~/components/ui/transition/BottomToUpTransitionView'
+import {
+  BottomToUpSoftScaleTransitionView,
+  BottomToUpTransitionView,
+} from '~/components/ui/transition'
 import { apiClient } from '~/lib/request'
 import { routeBuilder, Routes } from '~/lib/route-builder'
