+++ src/app/(app)/categories/[slug]/page.tsx
@@ -6,8 +6,10 @@ import { useParams } from 'next/navigation'
 
 import { BackToTopFAB } from '~/components/ui/fab'
 import { TimelineList } from '~/components/ui/list/TimelineList'
-import { BottomToUpSoftScaleTransitionView } from '~/components/ui/transition/BottomToUpSoftScaleTransitionView'
-import { BottomToUpTransitionView } from '~/components/ui/transition/BottomToUpTransitionView'
+import {
+  BottomToUpSoftScaleTransitionView,
+  BottomToUpTransitionView,
+} from '~/components/ui/transition'
 import { routeBuilder, Routes } from '~/lib/route-builder'
 
 import { getPageBySlugQuery } from './query'
