+++ src/app/(app)/friends/page.tsx
@@ -18,7 +18,7 @@ import { BackToTopFAB } from '~/components/ui/fab'
 import { Form, FormInput } from '~/components/ui/form'
 import { Loading } from '~/components/ui/loading'
 import { useModalStack } from '~/components/ui/modal'
-import { BottomToUpTransitionView } from '~/components/ui/transition/BottomToUpTransitionView'
+import { BottomToUpTransitionView } from '~/components/ui/transition'
 import { shuffle } from '~/lib/lodash'
 import { apiClient, getErrorMessageFromRequestError } from '~/lib/request'
 import { toast } from '~/lib/toast'
