+++ src/app/(app)/posts/page.tsx
@@ -7,7 +7,7 @@ import { PostPagination } from '~/components/modules/post/PostPagination'
 import { NothingFound } from '~/components/modules/shared/NothingFound'
 import { SearchFAB } from '~/components/modules/shared/SearchFAB'
 import { BackToTopFAB } from '~/components/ui/fab'
-import { BottomToUpTransitionView } from '~/components/ui/transition/BottomToUpTransitionView'
+import { BottomToUpTransitionView } from '~/components/ui/transition'
 import { OnlyDesktop } from '~/components/ui/viewport'
 import { apiClient } from '~/lib/request'
