+++ src/app/(app)/projects/page.tsx
@@ -6,7 +6,7 @@ import { GitHubBrandIcon } from '~/components/icons/platform/GitHubBrandIcon'
 import { ProjectList } from '~/components/modules/project/ProjectList'
 import { NothingFound } from '~/components/modules/shared/NothingFound'
 import { Loading } from '~/components/ui/loading'
-import { BottomToUpTransitionView } from '~/components/ui/transition/BottomToUpTransitionView'
+import { BottomToUpTransitionView } from '~/components/ui/transition'
 import { noopArr } from '~/lib/noop'
 import { apiClient } from '~/lib/request'
 import { useAggregationSelector } from '~/providers/root/aggregation-data-provider'
