--- src/app/(app)/says/query.ts
@@ -1,1 +1,0 @@
-export const sayQueryKey = ['says']
