+++ src/socket/handler.ts
@@ -12,7 +12,6 @@ import type { InfiniteData } from '@tanstack/react-query'
 import type { ActivityPresence } from '~/models/activity'
 import type { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime'
 
-import { sayQueryKey } from '~/app/(app)/says/query'
 import { setOnlineCount } from '~/atoms'
 import {
   deleteActivityPresence,
@@ -25,6 +24,7 @@ import {
   IcTwotoneSignpost,
   MdiLightbulbOn20,
 } from '~/components/icons/menu-collection'
+import { sayQueryKey } from '~/components/modules/say/hooks'
 import { DOMCustomEvents } from '~/constants/event'
 import { TrackerAction } from '~/constants/tracker'
 import { isDev } from '~/lib/env'
